"Update app to handle firebase cloud messaging and update image imports"
This commit includes updates to the app to handle firebase cloud messaging, including setting up a background message handler and a killed mode handler in the index.js file. Additionally, there are updates to the App.js file to include code for handling foreground notifications and getting the FCM token. Lastly, there are updates to the Images.js file in the Themes folder to update the import path for the Logo1 image.